added a deletelayers selector
